+++ src/Listeners/LoginListener.php
@@ -16,13 +16,17 @@ class LoginListener
         $this->request = $request;
     }
 
-    public function handle(Login $event): void
+    public function handle($event): void
     {
+        if (! $event instanceof (config('authentication-log.events.login') ?? Login::class)) {
+            return;
+        }
+
         if ($event->user) {
             $user = $event->user;
             $ip = $this->request->ip();
             $userAgent = $this->request->userAgent();
-            $known = $user->authentications()->whereIpAddress($ip)->whereUserAgent($userAgent)->first();
+            $known = $user->authentications()->whereIpAddress($ip)->whereUserAgent($userAgent)->whereLoginSuccessful(true)->first();
             $newUser = Carbon::parse($user->{$user->getCreatedAtColumn()})->diffInMinutes(Carbon::now()) < 1;
 
             $log = $user->authentications()->create([